js 쪼개기
modal, draw, 무한스크롤 각자 분리
