+++ src/main/webapp/resources/css/common/header.css
@@ -7,8 +7,9 @@ header {
     border: 1px solid rebeccapurple;
     box-sizing: border-box;
     height: 150px;
-    max-width: 100%;
-    position: sticky;
+    max-width: 500px;
+    width: 100%;
+    position: fixed;
     top: 0;
     z-index: 1;
     background-color: #ccc8c8;
